Initial domain event registry support.
Adds the DomainEvents service class which enables the registration
of callbacks for domain event classes.  It also provides the ability
to clear the registry of all callback, if necessary.

And finally, domain events can be "raised"; enabling all callbacks
to be called with the event provided to the callback block.